remove unnecessary explicit Traverser
