+++ metals/src/main/scala/scala/meta/internal/metals/FoldingRangeProvider.scala
@@ -3,15 +3,13 @@ package scala.meta.internal.metals
 import java.util
 import org.eclipse.lsp4j.FoldingRange
 import org.eclipse.lsp4j.FoldingRangeCapabilities
-import scala.meta.Term.Block
+import scala.meta.Template
+import scala.meta.Term
+import scala.meta.Tree
+import scala.meta._
 import scala.meta.internal.metals.FoldingMode._
-import scala.meta.internal.metals.FoldingRangeKind.Factory
 import scala.meta.internal.metals.FoldingRangeKind.Region
 import scala.meta.io.AbsolutePath
-import scala.meta.transversers.SimpleTraverser
-import scala.meta.Position
-import scala.meta.Template
-import scala.meta.Tree
 
 final class FoldingRangeProvider(val trees: Trees, mode: FoldingMode) {
   def getRangedFor(path: AbsolutePath): util.List[FoldingRange] = {
@@ -22,34 +20,14 @@ final class FoldingRangeProvider(val trees: Trees, mode: FoldingMode) {
   }
 
   private def findFoldingRanges(tree: Tree): util.List[FoldingRange] = {
-    val ranges = new FoldingRanges
-    new RangeFinder(ranges).apply(tree)
+    val ranges = new FoldingRanges(mode)
+
+    tree traverse {
+      case block: Term.Block => ranges.add(Region, block.pos)
+      case template: Template => ranges.add(Region, template.pos)
+    }
+
     ranges.get
-  }
-
-  private class RangeFinder(ranges: FoldingRanges) extends SimpleTraverser {
-
-    private def fold(tree: Tree): Unit = tree match {
-      case _: Block => createRange(Region, tree.pos)
-      case _: Template => createRange(Region, tree.pos)
-      case _ =>
-    }
-
-    override def apply(tree: Tree): Unit = {
-      fold(tree)
-      super.apply(tree)
-    }
-
-    private def createRange(factory: Factory, pos: Position): Unit = {
-      val range = factory.create(
-        pos.startLine,
-        pos.startColumn,
-        pos.endLine,
-        pos.endColumn
-      )(mode)
-
-      ranges.add(range)
-    }
   }
 }
